+++ NEWS.md
@@ -5,6 +5,7 @@
         of `terra::map.pal()`, that is the default palette for `terra::plot()`
         (`> 1.7.78`).
     -   `autoplot.SpatRaster()` now uses also `grass_db` as the default palette.
+    -   Add `scale_fill_princess_c()` scale family.
 -   Fix tests.
 
 # tidyterra 0.6.0